requirements.txt
Django==4.2.7
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
